changed from uint256 to uint16

+++ contracts/storage.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.0;
 contract RacerRegistry {
     struct Rider {
         string name;
-        uint256 bikeNumber;
+        uint16 bikeNumber;
     }
 
     mapping(address => Rider) public riders;
@@ -14,7 +14,7 @@ contract RacerRegistry {
         riders[msg.sender].name = _name;
     }
 
-    function racerBikeNumber(uint256 _bikeNumber) public {
+    function racerBikeNumber(uint16 _bikeNumber) public {
         riders[msg.sender].bikeNumber = _bikeNumber;
     }
 
@@ -22,7 +22,7 @@ contract RacerRegistry {
         return riders[msg.sender].name; // 
     }
 
-    function getBikeNumber() public view returns (uint256) {
+    function getBikeNumber() public view returns (uint16) {
         return riders[msg.sender].bikeNumber;
     }
 }
